+++ controller/userController.js
@@ -4,6 +4,7 @@ const nodemailer = require('nodemailer');
 const bcrypt = require('bcryptjs');
 const UserOTPVerification = require('../models/userVerificationModel');
 const hbs = require('nodemailer-express-handlebars');
+const { sendOtpVerificationEmail } = require('./signInController');
 
 //Fetch All User Function
 exports.fetchAllUsers = async(req,res) =>{
@@ -25,7 +26,8 @@ exports.registerUser = async(req,res) => {
         phoneNumber : req.body.phoneNumber,
         address : req.body.address,
         busNumber : req.body.busNumber,
-        scholarNumber : req.body.scholarNumber,   
+        scholarNumber : req.body.scholarNumber,  
+        verified:false, 
     });
     try {
         const userExistEmail = await User.findOne({email:req.body.email});
@@ -38,7 +40,7 @@ exports.registerUser = async(req,res) => {
            return res.json( 'Enrollment Number Already Exist'); 
         }
         await user.save().then((result) =>{
-           this.sendOtpVerificationEmail(result,res);
+           sendOtpVerificationEmail(result,res);
         }); //SAVE COMMAND
       
     } catch (error) {
@@ -46,56 +48,7 @@ exports.registerUser = async(req,res) => {
     }
 }
 
-//NODEMAILER 
-let transporter = nodemailer.createTransport({
-    host: "smtp.gmail.com",
-    auth : {
-        user: "mahobiashubham4@gmail.com",
-        pass: "hutufymgqfwvnzdh",
-    },
-});
-
-transporter.use('compile',hbs({
-    viewEngine: 'express-handlebars',
-    viewPath: './views/'
-}));
 
 
 
-exports.sendOtpVerificationEmail = async ({_id,email,fullName},res) => {
-    try {
-        const otp = `${Math.floor(1000 + Math.random() * 9000)}`;
-        const mailOption = {
-            from: "mahobiashubham4@gmail.com",
-            to: email,
-            subject: "Verify Your Email",
-            template:'verifyOTP',
-            context:{
-                fullName:fullName,
-                otp: otp,
-            }
-        };
-        transporter.close();
 
-        //Hashing the OTP - 
-       const hashOTP = await bcrypt.hash(otp,12); 
-       const newOtpVerification = await new UserOTPVerification({
-        userId : _id,
-        otp : hashOTP,
-        createdAt: Date.now(),
-        expiresAt : Date.now() + 3600000,
-       });
-
-       await newOtpVerification.save();
-       transporter.sendMail(mailOption);
-        return res.json({
-        message: "Verification otp Mail sent",
-        data:{
-            userId : _id,
-            email,
-        },
-       });
-    } catch (error) {
-        res.status(400).json({ success: false, message: error.message });
-    }
-}
