+++ models/userModel.js
@@ -37,6 +37,10 @@ const userSchema = new mongoose.Schema({
         type:String,
         required:true,
     },
+    verified:{
+        type:Boolean,
+        required:false,
+    },
 
 },
 {
